refactor: move RoundScreen components to App
