+++ src/screens/RoundScreen.tsx
@@ -1,28 +1,8 @@
-import { FileLock2 as WakeLockIcon } from '@styled-icons/bootstrap/FileLock2';
-import { Globe as LanguageIcon } from '@styled-icons/bootstrap/Globe';
-import React, { useState } from 'react';
-import { useTranslation } from 'react-i18next';
+import { useState } from 'react';
 import styled from 'styled-components';
-import HeaderButton from '../components/HeaderButton';
-import InstructionsModal from '../components/InstructionsModal';
-import LanguageModal from '../components/LanguageModal';
 import RoundDirectionArrow from '../components/RoundDirectionArrow';
-import Screen from '../components/Screen';
-import WakeLockModal from '../components/WakeLockModal';
-import strings from '../config/strings';
-import logger from '../utils/logger';
-import wakeLock from '../utils/wakeLock';
 
-const Header = styled.div`
-  width: 100%;
-  height: 48px;
-  padding: 0 4px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const RoundDirectionArrowWrapper = styled.div`
+const Wrapper = styled.div`
   flex-grow: 1;
   width: 100%;
   display: flex;
@@ -30,92 +10,23 @@ const RoundDirectionArrowWrapper = styled.div`
   justify-content: center;
   align-items: center;
 
-  // Solves the overflow issue when clicking the 'Learn more' button
+  // Prevents the user to click the arrow outside this wrapper
   overflow: hidden;
 `;
 
-// TODO: Test state & actions
-// TODO: Test round direction arrow props
-// TODO: Test instructions modal
-// TODO: Test language modal
-const RoundScreen: React.FunctionComponent = () => {
+function RoundScreen() {
   // Store direction value
   //  * Left = false
   //  * Right = true (default)
   const [roundDirection, setRoundDirection] = useState(true);
-  const [isInstructionsModalOpen, setIsInstructionsModalOpen] = useState(true);
-  const [isLanguageModalOpen, setIsLanguageModalOpen] = useState(false);
-  const [isWakeLockModalOpen, setIsWakeLockModalOpen] = useState(false);
-  
+
   // Actions
-  const closeInstructionsModal = (): void => setIsInstructionsModalOpen(false);
-
-  const closeLanguageModal = (): void => setIsLanguageModalOpen(false);
-  
-  const closeWakeLockModal = (): void => setIsWakeLockModalOpen(false);
-
-  const openLanguageModal = (): void => setIsLanguageModalOpen(true);
-
-  const openWakeLockModal = (): void => setIsWakeLockModalOpen(true);
-
-  // TODO: Move to wakeLock.ts to delegate responsability
-  const enableWakeLock = async (): Promise<void> => {
-    try {
-      logger.logInfo('[WakeLock] Locking screen...');
-      await wakeLock.enable();
-      logger.logInfo('[WakeLock] Screen locked!');
-    } catch (err) {
-      logger.logError('[WakeLock] An error occurred while locking the screen.', err);
-    }
-  };
-
-  const start = async (): Promise<void> => {
-    await enableWakeLock();
-    closeInstructionsModal();
-  };
-
-  const canToggleDirection = (): boolean => (
-    !isInstructionsModalOpen &&
-    !isLanguageModalOpen &&
-    !isWakeLockModalOpen
-  );
-
-  const toggleDirection = (): void => {
-    // The logic should be validated anyways besides the fact that the 
-    // user can't click on the screen if any modal is open
-    if (canToggleDirection()) {
-      setRoundDirection(!roundDirection);
-    }
-  };
-
-  // Texts
-  const { t, i18n } = useTranslation();
-  const wakeLockStatusText = t(wakeLock.isEnabled ? strings.WAKE_LOCK_STATUS_ENABLED : strings.WAKE_LOCK_STATUS_DISABLED);
-  const languageShortText = i18n.language.substr(0, 2).toUpperCase();
+  const toggleDirection = () => setRoundDirection(!roundDirection);
 
   return (
-    <React.Fragment>
-      <Screen data-testid="round-screen">
-        <Header>
-          <HeaderButton
-            icon={<WakeLockIcon size="20" title="Wake Lock" />}
-            text={wakeLockStatusText}
-            onClick={openWakeLockModal}
-          />
-          <HeaderButton
-            icon={<LanguageIcon size="16" title="Language" />}
-            text={languageShortText}
-            onClick={openLanguageModal}
-          />
-        </Header>
-        <RoundDirectionArrowWrapper onClick={toggleDirection}>
-          <RoundDirectionArrow direction={roundDirection} />
-        </RoundDirectionArrowWrapper>
-      </Screen>
-      <InstructionsModal isOpen={isInstructionsModalOpen} onClose={start} />
-      <LanguageModal isOpen={isLanguageModalOpen} onClose={closeLanguageModal} />
-      <WakeLockModal isOpen={isWakeLockModalOpen} onClose={closeWakeLockModal} />
-    </React.Fragment>
+    <Wrapper onClick={toggleDirection}>
+      <RoundDirectionArrow direction={roundDirection} />
+    </Wrapper>
   );
 };
